Fixed server action (Security)

+++ actions/delete-s3-file.ts
@@ -1,12 +1,19 @@
 "use server";
+import { auth } from "@clerk/nextjs/dist/types/server";
 import { UTApi } from "uploadthing/server";
 
 const utapi = new UTApi();
 
-export const deleteS3Files = async (key: string | string[]) => {
-  try {
-    await utapi.deleteFiles(key);
-  } catch (error) {
-    console.log(error);
+export const deleteS3Files = async ({
+  key,
+  user,
+}: {
+  key: string | string[];
+  user: string;
+}) => {
+  const { userId } = auth();
+  if (!userId || userId !== user) {
+    throw new Error("Unauthorized");
   }
+  await utapi.deleteFiles(key);
 };
